requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3

openai==1.61.0
google-cloud-texttospeech>=2.16.5,<3.0.0

